requirements.txt
altair==5.1.2
appdirs==1.4.4
appnope==0.1.3
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.2
bokeh==3.3.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
contourpy==1.1.1
cryptography==41.0.7
cycler==0.11.0
debugpy==1.6.7.post1
decorator==4.4.2
dnspython==2.4.2
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
Faker==20.1.0
favicon==0.7.0
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
Flask-PyMongo==2.3.0
fonttools==4.42.1
frozendict==2.3.8
geographiclib==2.0
geopy==2.4.0
gitdb==4.0.11
GitPython==3.1.40
htbuilder==0.6.2
html5lib==1.1
icecream==2.1.3
idna==3.4
imageio==2.33.0
imageio-ffmpeg==0.4.9
importlib-metadata==6.8.0
iniconfig==2.0.0
ipykernel==6.25.1
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.5
lml==0.1.0
lxml==4.9.3
Markdown==3.5.1
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==10.1.0
moviepy==1.0.3
mpmath==1.3.0
multitasking==0.0.11
nest-asyncio==1.5.7
numpy==1.26.0
openpyxl==3.1.2
packaging==23.1
pandas==2.1.0
parso==0.8.3
peewee==3.16.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.10.0
plotly==5.18.0
pluggy==1.3.0
proglog==0.1.10
prompt-toolkit==3.0.39
protobuf==4.25.0
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.0
pycparser==2.21
pydeck==0.8.1b0
pyexcel==0.7.0
pyexcel-ezodf==0.3.4
pyexcel-io==0.6.6
pyexcel-ods3==0.6.1
pyexcel-xlsxw==0.6.1
pygame==2.5.2
Pygments==2.16.1
PyJWT==2.8.0
pymdown-extensions==10.5
pymongo==4.5.0
pyparsing==3.1.1
PySide6==6.5.2
PySide6-Addons==6.5.2
PySide6-Essentials==6.5.2
pytest==7.4.2
python-dateutil==2.8.2
python-http-client==3.3.7
pytube==15.0.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
resend==0.6.0
rich==13.6.0
rpds-py==0.12.0
scipy==1.11.3
seaborn==0.13.0
sendgrid==6.10.0
shiboken6==6.5.2
six==1.16.0
smmap==5.0.1
soupsieve==2.5
st-annotated-text==4.0.1
stack-data==0.6.2
starkbank-ecdsa==2.2.0
streamlit==1.29.0
streamlit-bokeh-events==0.1.2
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.61
streamlit-embedcode==0.1.2
streamlit-extras==0.3.5
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
sympy==1.12
tenacity==8.2.3
texttable==1.7.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
unicode==2.9
Unidecode==1.3.7
urllib3==2.0.5
validators==0.22.0
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.3.7
XlsxWriter==3.1.6
xlwt==1.3.0
xyzservices==2023.10.1
yfinance==0.2.29
zipp==3.17.0
